+++ tests/test_fornax_jupyter.py
@@ -43,9 +43,8 @@ def test_notebooks_folder():
     assert os.path.exists(f'{notebook_dir}/heasarc-tutorials')
     assert os.path.exists(f'{notebook_dir}/mast-tutorials')
 
-# commented out because we need endir for the python binary
-# def test_env_dir_not_exist():
-#     assert not os.path.exists(os.environ['ENV_DIR'])
+def test_env_dir_not_exist():
+    assert not os.path.exists(os.environ['ENV_DIR'])
 
 
 def test_support_data_synlink():